+++ PySoap2_gpu/layers/c_code/polynomial_transformation_2d.py
@@ -1,15 +1,15 @@
 """ P1[i, :] is assumed to be the i-th polynomial """
 
 polynomial_2d_transform_c_code = """
-__kernel void polynomial_transform_2d(__global const float *P1, __global const float *P2, 
-                                      __global const float *images, const int M1, const int M2, 
-                                      const int input_length, __global float *out)
+__kernel void polynomial_transform_2d(__global const double *P1, __global const double *P2, 
+                                      __global const double *images, const int M1, const int M2, 
+                                      const int input_length, __global double *out)
 {
     int n = get_global_id(0);
     int i = get_global_id(1);
     int j = get_global_id(2);
 
-    float total = 0.0;
+    double total = 0.0;
     for (int x = 0; x < M1; x++)
     {
         for (int y = 0; y < M2; y++)
@@ -20,9 +20,9 @@ __kernel void polynomial_transform_2d(__global const float *P1, __global const f
     out[n*input_length + i*M2 + j] = total;
 }
 
-__kernel void polynomial_transform_2d_multi(__global const float *P1, __global const float *P2, 
-                                            __global const float *images, const int M1, const int M2,
-                                            const int M3, const int input_length, __global float *out)
+__kernel void polynomial_transform_2d_multi(__global const double *P1, __global const double *P2, 
+                                            __global const double *images, const int M1, const int M2,
+                                            const int M3, const int input_length, __global double *out)
 {
 
     int n = get_global_id(0);
@@ -32,7 +32,7 @@ __kernel void polynomial_transform_2d_multi(__global const float *P1, __global c
     int j = J / M3;
     int l = J % M3;
 
-    float total = 0.0;
+    double total = 0.0;
 
     for (int x = 0; x < M1; x++)
     {
